requirements.txt
# Telegram Bot
aiogram==3.4.1
python-telegram-bot==20.7

# YouTube Download
yt-dlp==2024.1.7

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache & Queue
redis==5.0.1
celery==5.3.4

# File Storage
boto3==1.34.34
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1 